Kleine code refactor uitgevoert
