+++ yaeger.spacesimulator/src/main/java/com/yaeger/spacesimulator/data/ObjectPlacementData.java
@@ -1,6 +1,7 @@
 package com.yaeger.spacesimulator.data;
 
 import com.github.hanyaeger.api.Coordinate2D;
+import com.yaeger.spacesimulator.services.ConfigService;
 
 import javafx.scene.paint.Color;
 
@@ -13,7 +14,10 @@ public class ObjectPlacementData {
 	public ObjectPlacementData() {
 		this.startPosition = new Coordinate2D();
 		this.stopPosition = new Coordinate2D();
-		this.color = new Color(1, 1, 1, 1);
+		
+		this.volume = Integer.parseInt(ConfigService.getValue("base-planet-volume"));
+		this.density = Integer.parseInt(ConfigService.getValue("base-planet-density"));
+		this.color = Color.web(ConfigService.getValue("base-planet-color"));
 	}
 	
 	public void reset() {
