+++ yaeger.spacesimulator/src/main/java/com/yaeger/spacesimulator/services/ConfigService.java
@@ -1,4 +1,4 @@
-package com.yaeger.spacesimulator.service;
+package com.yaeger.spacesimulator.services;
 
 import java.io.File;
 import java.io.FileNotFoundException;
