+++ yaeger.spacesimulator/src/main/resources/config.properties
@@ -2,3 +2,10 @@
 font-family = monospace
 font-size-m = 16
 font-size-s = 14
+
+#Base planet values
+base-planet-volume = 500
+base-planet-density = 25
+base-planet-color = 0xF1F1F1
+
+gravity-constant = 0.1D
